pyproject.toml
[project]
name = "prompt-lab"
version = "0.1.0"
description = "Prompt refinement harness for DBT chatbot (AWS Bedrock)"
authors = [{ name = "Priyanshi Garg" }]
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.34",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[tool.uv]
dev-dependencies = [
    "black",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/prompt_lab"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

